requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.0
streamlit>=1.30.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=23.10.0
ruff>=0.1.4
aiohttp>=3.9.0
openai>=1.3.0  # For OpenAI provider
ollama>=0.1.10  # For local Ollama provider
sumy>=0.11.0  # For text summarization
nltk>=3.8.1  # Required for sumy
